created map folder, component and base styling. adjusted routes to reflect change to home screen with map rendering
